Update and rename StrategyStateImpl.java to StrategyStateFactoryImpl.java
